Specify an ext subpath in the test/format
Instead of baking the in the full system path to the extensions,
just allow CMake to specify whether the extension lives under a
specific subdirectory (i.e. .libs on autoconf). We lose a lot of
portability if we hardcode a full absolute path in the test/format,
i.e. the binary becomes invalid if executed under a different
sysroot.