Log library verbose messages along with format trace messages.
Use the library verbose functions to create the message prefix for format's tracing.

Allow reconfiguration of verbose options, the bug is that any reconfiguration of the
connection clears the existing verbose configurations, regardless of whether or not
the configuration includes verbose settings.

Rework the general error functions to handle big error messages, necessary to avoid
truncating long verbose messages.

Remove the option to trace format operations in the primary database, it's no longer
needed now that primary database verbose messages are logged in the secondary.

Revert the change to move trace options into the CONFIG file, it's more trouble than
it's worth.