WT-8424 Use consistent toolchain in little-endian
When compiling WiredTiger in the Evergreen little-endian variant,
the build is configured with the v3 mongodbtoolchain. We
subsequently need to ensure the PATH is set to point to the mongodb
toolchain when running make. Otherwise build scripts invoked
during the compiliation process i.e. workgen's 'setup.py', will
be called with a mis-matched toolchain.