WT-8395 Save connection write generation number along with checkpoint snapshot
To identify the saved checkpoint snapshot in the metadata file during recovery,
WiredTiger calculates the connection level write generation number from all the
btree checkpoints and compares it against the saved write generation number in
the checkpoint snapshot before proceeding with RTS based on that checkpoint
snapshot.

The versions where the connection write generation number is not saved, during
the restart, the comparison fails and RTS doesn't perform using the checkpoint
snapshot.