WT-8420 FLCS support for test_checkpoint (#7209)
* Add FLCS support to test_checkpoint.

This involves (a) adding logic in various places to deal with FLCS not
having deleted values, and (b) coming up with a scheme such that the
values in other tables can be compared to FLCS tables without either
being vacuous or requiring more than 8 bits of storage.

For the latter, what I came up with is to store the offset of the
first non-prime digit, with a reserved value for "none" and another
reserved value for "we lost track, can't crosscheck this key". The
latter doesn't happen all that often. (XXX: should rerun the check for
that now that other things have been fixed.) Pick an encoding that
also allows never storing zero ourselves.

So in turn the former just requires adding code in a few places that
skips over zero values as if they're deleted, since the only way zeros
appear is via deletion.

Also:

 - Change the test lists to use 4 mixed instead of 3 mixed tables when
doing mixed tables, because there are now four table types. Sort the
test lists by table type and add some headings to make things easier
to follow.

 - When verifying, pick the first table that isn't FLCS and isn't LSM
to be the reference for comparison. (If all the tables are FLCS or
LSM, use the first table as before.) Currently this has no effect as
if you don't choose mixed tables the tables are all the same, and if
you do, the ordering of types is hardcoded with VLCS first, but I
could imagine that changing to allow specifying particular
combinations of table types, and this way it will continue to work as
intended instead of silently using a weaker comparison when an FLCS
table is first.

* Sort the arguments in the test list to put the common ones first.

Particularly in the shell version (where they now all line up) this
makes them a lot easier to read.

No functional change.

* Fix clang warnings.

Also I apparently forgot to run s_all, though I would have sworn I
had... add "flcs" as a spelling word along with "FLCS" so we can have
functions named flcs_*.

* Fix slipup in smoke.sh.