Move the rest of the block manager's  mmap code into the block-cache code (which implies pre-loaded blocks will now be cached in the block cache after being read into the WiredTiger cache).
This commit removes preload's call to handle->fh_advise() with the WT_FILE_HANDLE_WILLNEED flag.
On Linux that call uses posix_fadvise(2) to advise the OS a block will be used in the near future
(the call "initiates  a nonblocking read of the specified region into the page cache"). Because
the pre-load implementation falls back to a blocking read(2) if the file isn't mapped into
memory, it's unclear it's useful to make that additional system call, and a quick smoke-test
didn't show me any performance difference between code with or without the call.