WT-8257 Verbose Messaging Doc Page (#7241)
* Added doc page on verbose messaging

Updated the doxygen documentation with a new page on WiredTiger's
verbose messaging system. This reflecting the recent changes to the
configuration interface, including a discussion about the verbosity
levels and categories.

* Create a verbose messaging C example

To demonstrate basic usage of the verbose messaging API, created
a small example C program. The program configures a wiredtiger
connection with verbose components. This additionally being useful
to provide code snippets that can be used in an accompanying doc
page.