WT-8396 Integrate '__wt_progress' to support JSON-encoded messages (#7220)
* New json_output option: progress

Introducing a new option for the json_output setting: progress. This
will be used for messages that are generated from the __wt_progress API.

* Creation of a new verbose category: WT_VERB_PROGRESS

* Generate a more detailed message in __wt_progress

In order to benefit from the json_output setting in __wt_progress, we
reuse __eventv_gen_msg to generate a more detailed message.
Using __eventv_gen_msg also gives us the ability to have a JSON
or flat string format output for progress-related messages.

* Add a new input argument to __wt_progress that represents a category

As __wt_progress can be called from different areas in the code, add a
new category argument to __wt_progress in order to specify the component
that is making progress. New categories had to be created with the suffix
"PROGRESS" for clarity.

* Use of WT_VERBOSE_NOTICE

Messages generated through the __wt_progress API use the verbosity level
WT_VERBOSE_NOTICE.