Factor "when removing a key due to a tombstone..." out of leaf reconcilation.
Move the code (that there's been three copies of) to rec_write.c; move
the state to WT_RECONCILE.

This reduces the number of history cursors created from up to one per
page to no more than one per WT_RECONCILE, which is also a minor
improvement. The new logic resets the cursor after each page once it's
created, so as not to accidentally sit on any history store pages, but
there seems to be no reason to destroy and recreate it from one page
to the next.

This does, however, mean that __rec_destroy can now fail, even if the
failure path is probably theoretical.