WT-8485 Revert JSON messages for wt_progress (#7259)
Upon further evaluation, the purpose of the progress interface
was misinterpreted, where the message passed via the
interface should just be an operation string. Hence there is no
requirement for this string to be a complex message string.
This reverts changes introduced in WT-8396. However manually
reverts changes only relating to '__wt_progress' (as there are some
other general improvements worth keeping in the original commit).