WT-8395 Allow clean shutdown upgrade only from version 10.0.0 (#7239)
WiredTiger version 10.0.0 can possibly write inconsistent data to disk on a checkpoint. To fix the
inconsistent data, WiredTiger started saving a checkpoint snapshot to fix this problem to fix it upon restart.
In version 10.0.0, a stale version of the checkpoint snapshot is saved and it is corrected in version 10.0.1.
Using the stale version of the checkpoint snapshot can lead to data corruption issues. To avoid data corruption
issues,  allow unclean shutdown upgrade from only 10.0.1 onwards and error out upgrading the version
on an unclean shutdown from the 10.0.0 version.