Change history cursors to use per table history files
Not completely working.  Some python tests hang.

Changes:
* __wt_curhs_open() uses its session dhandle to determine which history
file to open a cursor on.

* Disable __wt_hs_config().  Instead set WT_DHANDLE_HS on history files
when we allocate the dhandle in __wt_conn_dhandle_alloc().  Set the
WT_BTREE_NO_LOGGING flag on history btrees in __wt_btree_open()

* One branch through __rollback_to_stable_btree_apply() winds up using
a history cursor without an session dhandle.  (It gets the btree ID
it needs by reading the metadata.)  I updated this to explictly open
the dhandle.

* Changed __wt_open_cursor() to allow opening a metadata cursor while a
session has a history cursor.  This is needed because __wt_curhs_open()
bumps the session's count of history cursors before it does the real
work, which can now include reading the metadata for the desired history
file.  In retrospect it would probably be cleaner to change where that
count is incremented, rather than changing the ASSERT in
__wt_open_cursor().

* Where code needs to look at a uri and decide if it is a history file,
we now just look at the file prefix (WiredTigerHS), rather than matching
the full uri.  I only changed this in places where I ran into issues.
Need to check for other occurances in the code.