Reorder csuite tests in the long-test task
In Evergreen, variables live within the scope of a task. If a function
defines a variable once, it will persist and be visible to the
subsequent function calls within that task. If a test should not use
extra variables, it should be executed before the extra variables are
defined.

(cherry picked from commit fcbf283e0675f8e45c9f22dc1c11610b3bef8902)