Fix bug exposed by the original change.
Store the auxiliary data start in the WT_REC_CHUNK and use that to
write the auxiliary header. Using the value from the WT_RECONCILE
(which applies to the current chunk) blows up when writing the
previous chunk if the two aren't the same size, which they often are
but in general are not, especially but not only with the original
change.