UBSAN fix: If a caller tries to assign an empty string to a WT_ITEM, we can end up not allocating any memory at all, and then setting an offset (of 0), from a NULL pointer. Write up the case and make sure it doesn't happen.
For the record, the other approach would be to always allocate at least a few bytes for the
WT_ITEM, but that approach leads to repeated failures to free memory from a variety of sources,
for example, checkpoints. Historically, callers of this function know we won't actually allocate
memory if we don't need it, and it seems like a lot more work to fix it than to leave it alone.