WT-8479 Expose verbose category identifiers in public WiredTiger API (#7257)
* Expose verbose categories and levels
This commit moves the definitions of verbose categories and levels from
verbose.h to wiredtiger.in in order to expose them publicly.

* Script to generate the verbose category strings
This commit adds a new script: verbose.py. Based on tags, it generates
the string values of each verbose category defined in wiredtiger.in.

* Rework verbose str macro to avoid redeclaration
Update the verbose category string macro to just generate an array of
strings that can be statically declared once in err.c.

* Add more checks in validate_json_schema
Now we have exposed the verbose categories to the Python API as well, we
can update our python tests to check that the categories present in the WT
messages match their definitions.

* Refined documentation on public verbose categories
Since verbose categories are subject to change over time, updated
the public API documentation to only document the verbose
categories that are likely to remain constant. Additionally added
a small sentence indicating the volatile nature of these
definitions.

Co-authored-by: Alison Felizzi <alison.felizzi@mongodb.com>