WT-8534 Allow retrieving the checkpoint snapshot for backup restore recovery (#7309) (#7334)
WiredTiger versions 10.0.1 onward have a valid checkpoint snapshot on-disk. There was a bug
in some versions of WiredTiger that are tagged with the 10.0.0 release, which saved the wrong
checkpoint snapshot, so we ignore the snapshot when it was created with one of those versions.
Versions of WiredTiger prior to 10.0.0 never saved a checkpoint snapshot. Additionally the turtle
file doesn't always exist (for example, backup doesn't include the turtle file), so there isn't always
a WiredTiger version is available. If there is no version available, assume that the snapshot is valid,
otherwise restoring from a backup won't work.

Co-authored-by: Alexander Gorrod <alexander.gorrod@mongodb.com>
(cherry picked from commit 172d0033bd8608828bcb79cf935e458493c9aba1)