Restored diagnostics in configure_combinations.sh
Restored the original diagnostic flags in
'configure_combinations.sh' when compiling with gcc. A number of
these flags were reworked/removed since they are incompatible
with SWIG generated code. This is due limited coverage in the
Python API compiliation in CMake for GCC builds. Rather than rework
the original flags in configure_combinations, expanded the CMakeLists
for the Python API to ignore erroneous diagnostics, as this can
be easily triggered under '-Werror -Wall'.