Enable diagnostics for spinlock-gcc-test and spinlock-pthread-adaptive-test.
The new test_rollback_to_stable30.py tests expect a verbose transaction
dump, but this only happens if diagnostics are enabled. Turn these on during
the configure stage.

The alternative would've been to change the log message that the test
expects to see, but there doesn't seem to be any harm in enabling
diagnostics.