WT-8633 Use NEW behaviour for CMP0078 in CMake (#7395)
If available, let CMake use the NEW behaviour for the CMP0078
Swig policy. The OLD behavior for policy CMP0078 is intended to
be removed from a future version of CMake. This removing the
warning created at configuration time for newer versions of CMake
and future-proofing the Python SWIG compilation when the OLD
behaviour support for CMP0078 is removed.