WT-8616 Btree walk functions can ignore the caller's "skip function" (#7391)
Check if the caller wants to skip the page regardless of the outcome of previous "skip this page tests". If the caller wants to skip a class of pages (for example, compaction wants to skip all leaf pages), we have to check in the caller's skip function regardless of other tests.