WT-7662 Format timed out with prepare-conflict (#7421)
Periodically we're flagging a failure in a prepared transaction, and as far as we can tell, it's not real, it's just that prepared transactions are staying around without resolution for too long (the last one was a measured 70 seconds). Stop sleeping after prepared transactions and increase the amount of time we'll wait for prepare-conflict resolution from 1 to 2 minutes.

(cherry picked from commit b568512de4bb46ae4c6620fcf37ce9d3817e1d73)