WT-8570 Do not increase oldest ID during recovery
Increasing the oldest ID during the recovery can lead to a
scenario where the eviction threads that operation in recovery
can check the global visibility according to the current oldest ID
and it may go wrong until the recovery rollback to stable is finished.

Also let all the history store records return to the rollback to stable
irrespective of global visibility.