test: Rework fallback in format.sh script
We can no longer statically fallback to the 'build_posix'.
Since there's no static CMake build directory we can default to,
the format.sh will default to using the build directory the script
resides under. Our CMake build syncs a copy of the format.sh
script to the build directory. This default assumes the script
being executed is a copy that lives under the build directory.
Otherwise it depends on passing the build directory path, which
has been added as an addtional cmdline option.