test: Rework 'testutil_build_dir' helper
Reworked the 'testutil_build_dir' utility helper to require the
user to set the build from the command line options (if calling
the function). We unfortunately can't depend on a known/constant
build directory (the user could have multiple out-of-source build
directories). There's also not really any OS-agnostic mechanism
we can here use to discover the build directory the calling test
binary exists in. To keep it simple, we will just rely on the
command line option being set.