WT-6707 Allow eviction to operate with snapshot when checkpoint is running
Earlier eviction was not allowed to operate with transaction
snapshot when the checkpoint is running in parallel as the eviction
operating with transaction snapshot can lead to inconsistent
on-disk checkpoint information. RTS has the support of fixing
the inconsistent checkpoint with the recovered checkpoint snapshot.
Allowing eviction to operate with snapshot improves the cache usage.