Defined 'None' build mode
The 'None' build mode is the default CMake build mode, where no
additional compiler flags are appended. We want to explicitly
define it as newer versions of CMake will display warning messages
about the None build mode not existing in the CMake cache. This
change shouldn't have an effect on the build, with or without the
line (we still get the same functionality). This change is just
intended to suppress spurious warning messages.