WT-8704 Tidy up CMake warnings on Windows builds (#7452)
* Added CMAKE_MODULE_LINKER_FLAGS to build modes.
* Defined 'None' build mode
The 'None' build mode is the default CMake build mode, where no
additional compiler flags are appended. We want to explicitly
define it as newer versions of CMake will display warning messages
about the None build mode not existing in the CMake cache. This
change shouldn't have an effect on the build, with or without the
line. This change is just intended to address warning messages.
* Moved Windows warning diagnostics to main config
The MSVC warnings C4090 & C4996 are quite prominent irrespective
of a strict or non-strict build. As we don't intend to fix these
issues, moved these diagnostic warnings to the main Windows config
file to reduce warning logs.