WT-7598 Remove autoconf & SCons build systems  (#7454)
The CMake build system is to become the main driver for POSIX 
and Windows builds. We now want to deprecate support for the 
old build systems and remove them from future releases of the 
WiredTiger standalone repository. The changes here are 
three-fold:
* Re-works our various scripts and utilities to use a CMake build 
path. This mainly being scripts under 'dist/', 'test/' & 'bench/'
* Updated lingering documentation. This mainly removing doc 
and comment references to the autoconf or SCons build system.
* Remove autoconf and SCons. Deletes all sources and paths related
 to the two build systems.