WT-8705 Add a CMake config_choice to use a local version of the SDK (#7493)
Add a compiler flag (IMPORT_S3_SDK) to CMake to allow the user to specify how to import the AWS S3 SDK (i.e. whether to download as an ExternalProject or use a local installation). This can be optionally set alongside ENABLE_S3.

IMPORT_S3_SDK can be set to:

* external - which will prompt the AWS SDK to be downloaded and installed during the build step.
* package - CMake will instead search locally for an installation of the SDK to use.

When the ENABLE_S3 flag is passed and an option for IMPORT_S3_SDK has not been chosen, the default behaviour is to look for a local installation of the SDK (i.e. the package option).