WT-8679 Add support for arm64 Evergreen testing
Our current standalone evergreen CI setup doesn't include testing
on arm64. The MongoDB server supports arm64 hosts and testing for
arm64 builds of WT are done indirectly via MongoDB. There may be
parts of WT that are not current exercised by the server
regression testing, and it would be better to catch any arm
specific issues in WT itself. Since arm64 is becoming increasingly
important in the clould, we should perform testing on arm64 at a
scope that is equivalent to our other target architectures.