WT-8716 Configure S3 cache and Implement S3 filesystem's file-exists check (#7477)
Implemented the S3Connection:ObjectExists API. It is implemented using a HEAD request to S3 to obtain only the object's metadata and not the object itself. The S3 Cache has also been configured as a directory on disk. When checking whether a file exists; the extension first checks the cache, then S3.