Switch roundup_timestamps=(prepare) back to using oldest.
This is safe only during application recovery for preparing
transactions that had already prepared successfully before a crash.

Otherwise, prepare timestamps are still required to be >= stable.

(It would be nice to change the name of the config variable to make it
look less like a general-purpose facility and link it more closely to
what it's actually for, but that requires downstream coordination that
isn't worth the effort for the time being.)