WT-8362 Remove or rewrite HS entries of a key when OOO tombstone is written to datastore (#7520)
Remove or rewrite the historical versions in the history store of a key
whenever an out-of-order or mixed-mode tombstone is selected to
write to the data store irrespective of whether the tombstone is
globally visible or not. When the tombstone is not globally visible,
all the existing history store versions are reinserted using the OOO
tombstone value.