Merge the test fixups from the prior WT-8747 pull request.
Specifically: avoid reading between the commit and durable timestamps
of updates, except when we want to do it on purpose in durable_ts03
and hs06.

Reading between commit and durable is problematic and there's no need
to tickle those problems unless we mean to. (The RTS tests in
particular have other goals.)