WT_CURSOR.remove cleanups
Don't use the current WT_UPDATE type when checking for collisions in __curfile_update_check(),
there's no reason to assume the current update type is going to be replicated in the update
and if the current update is a tombstone, we trigger the "two tombstones in a row" check.

Rework WT_SESSION.truncate code: WT_NOTFOUND from WT_CURSOR.next means we've hit the end of
the tree and can return success. WT_NOTFOUND from the remove function means the item was
removed and doesn't exist, implying an update collision (the key was visible to us as we
walked the tree, but then a remove happened, so there's must have been a visibility problem),
return WT_ROLLBACK to describe that visibility problem.

If WT_CURSOR.update in a table fails to remove an index entry, ignore the error and continue.

If WT_CURSOR.remove in a table fails to find an index or table entry, don't treat WT_NOTFOUND
differently from any other error, it's a hard failure.

Split __wt_txn_modify_check() into two parts, with the previous visibility checks in one function,
and the remove's "two tombstones in a row" check and rolling back based on a debug configuration in
a wrapper function. Put the two latter tests AFTER checking for correctness so we don't prioritize
not-found errors over rollback errors, that is, if there's a visibility problem we return rollback,
and only if there's no visibility problem do we consider if we're removing a non-existent key.
Also, don't check the debug configuration until we've already checked for correctness, there are
useful asserts in the correctness checks.

Change test/checkpoint to include the function/line-number pair in failure messages.

Fix test/checkpoint and wt2323_join_visibility to handle WT_CURSOR.remove returning WT_NOTFOUND.

Update some comments to reflect the new WT_CURSOR.remove semantics.