WT-8703 Extend recovering from backup to allow partial restores (#7634)
Provide a functionality to perform partial restores from backup. The use case is that the user can copy over selected tables, along with passing in a target uri list. From WiredTiger's perspective, this means we must clean all the metadata entry references using the use of a hash table and truncate all the history entries to the tables that are not part of the database anymore.