WT-8963 Implement a stress test for insert operation (#7676)
The motivation for this test is to try and reproduce BF-24385 by stressing insert functionality. The test creates a lot of threads that concurrently insert a lot of records with random keys. Having a large memory page ensures that we have big insert lists. Big cache size allows having more dirty content in the memory before eviction kicks in.