Fix reading from old checkpoints.
Don't update checkpoint trees' base write generation from the current
connection; that causes transaction IDs to disappear and items that
should be invisible items to become visible.

Add more scenarios to checkpoint10 and 11 (as if there aren't enough
already) that reopen the connection before reading the checkpoint.
Also add scenarios for logged tables to checkpoint10 as that wasn't
being tested and was potentially affected by this change.

Thanks to Hari for noticing that there was a problem.