Restore the illusion that FLCS deleted values are zero, for remove.
The changes to avoid creating multiple sequential tombstones broke
this, so you couldn't delete a just-deleted value until the page got
reconciled so there wasn't an overt tombstone any more. Fix it by
adding a dummy zero value immediately before the tombstone if we're
sitting on a key that exists (compare == 0) but is not valid, which
for FLCS only happens for rows with a tombstone update.