WT-8362 Truncate HS when ooo or mm tombstone is written to datastore (v4.4 backport) (#7752)
* WT-8362 Remove or rewrite HS entries of a key when OOO tombstone is written to datastore (#7520)

Remove or rewrite the historical versions in the history store of a key
whenever an out-of-order or mixed-mode tombstone is selected to
write to the data store irrespective of whether the tombstone is
globally visible or not. When the tombstone is not globally visible,
all the existing history store versions are reinserted using the OOO
tombstone value.

(cherry picked from commit a0f947fb7a0a18b6707ae3e760d77aa332d58cfd)

* More work around the reconciliation api

* Disable FLCS scenario

* WT-8550 Bump the maximum number of opened hs cursors to 3 (#7362)

* WT-8550 Bump the max number of opened hs cursors to 3

In some scenarios, it is possible to open up to three hs cursors. This
commit increases the hard limit on the number of hs cursors that can be
opened at once.

Co-authored-by: Hari Babu Kommi <haribabu.kommi@mongodb.com>