WT-9096 Fix search near to return wrong key/value if key doesn't exist
Avoid using the cursor temporary key as the history store key when the
cursor search compare is not found an exact match or if the match is found
in the insert list. In the above described scenarios, the cursor tmp key
is not the same key where the cursor slot is pointing. Generate the key
from cursor slot in those scenarios to avoid checking the wrong key in
the history store for the visibility.