WT-9096 Fix search near returning wrong key/value sometimes when key doesn't exist (#7777)
Avoid using the cursor temporary key as the history store key when the
cursor search compare is not found an exact match. When cursor search
didn't find an exact match, the cursor tmp key and the cursor slot can diverge.
Generate the key from the cursor slot in those scenarios to avoid checking
the wrong key in the history store for visibility.