Don't skip trees on forced checkpoints.
There's a test in __checkpoint_apply_operation that results in only iterating over
trees that are open; it doesn't use or involve WT_BTREE_SKIP_CKPT so is easily missed.

It has overrides for when the checkpoint is named or when checkpoints are being
dropped; also add an override for when force is set.

Otherwise the recovery checkpoint can skip trees and this leads to various possible
thorny problems with write generations.

Problem does not manifest in any default tests, but does turn up with test_checkpoint
if you do more runs than we do by default.