Deal with mismatched write generations a different way.
- Include the write generation from the snapshot metadata in WT_CKPT_SNAPSHOT
and thus return it out when opening a checkpoint.

 - Stash this write generation in the checkpoint cursor.

 - During checkpoint cursor operations, store it in the session.

 - In the unpacking code, use this write generation in preference to
the btree's own base write generation if we are reading a checkpoint
and it's not zero. Assert that if it's different from the btree's base
write generation that it's newer.

(It can be zero if we opened a checkpoint that doesn't have the
information. In principle we could store the btree's own base write
generation instead and drop the zero test; but getting it out is a
nuisance.)