Fix a problem that appeared with the new test scenarios.
It seems that if you instantiate a fast-deleted page from a
checkpoint, then evict it, the instantiated page gets discarded, but
the fast-truncate info associated with the page is not regenerated
when the page is read back in. (That only happens when/if the internal
page pointing to the truncated leaf page is reread.) This causes the
tombstones on the instantiated page to mysteriously evaporate and
causes the truncated values to randomly come back to life depending on
the whim of the eviction schedule.

To fix this, don't discard the fast-delete information when
instantiating pages in read-only trees. (All read-only trees, not just
checkpoints, because this will behave the same way for an ordinary
read-only tree.)

Also, there are two visible_all checks in the fast-truncate path that
need to be short-circuited when reading a checkpoint, because
visible_all checks test the current state and not the checkpoint
state. For these we want to treat the fast-truncate as never globally
visible; the fast-truncate info got written out because it was visible
to someone at the time, and that won't ever change.

(There is an exception, which is that if the same btree-level
checkpoint gets reused for multiple unnamed (only) checkpoints, the
original global context it was created in is no longer accessible and
so the truncate might in fact be globally visible via all checkpoint
read paths. This is not worth writing or testing special-case code
for.)