WT-9183 Fast-truncate in readonly trees can result in truncated items reappearing (#7835)
If you instantiate a fast-deleted page from a checkpoint, then evict it, the instantiated page gets discarded, but
the fast-truncate info associated with the page is not regenerated when the page is read back in. (That only happens when/if the internal page pointing to the truncated leaf page is reread.) This causes the tombstones on the instantiated page to mysteriously evaporate and causes the truncated values to randomly come back to life depending on the whim of the eviction schedule.