Disable crosschecks in tests.
These tests use timing_stress_for_test=[checkpoint_slow] and a
background checkpoint Python-thread to create a torn transaction in a
checkpoint. This is important for testing that the torn transaction
isn't visible. Actually getting one relies on racing with the
background thread, and (especially under load) sometimes one loses the
race and the checkpoint isn't inconsistent. When I wrote the tests I
added code to crosscheck that this doesn't happen, because when it
does the test succeeds but doesn't actually test anything interesting.
However, it seems that we do get occasional failures on the test
machines, which are at best a nuisance and at worst can get mistaken
for real problems.

Consequently, I'm going to disable the crosschecks until we have a
reliable way to generate inconsistent checkpoints. The current method
waits for the checkpoint to start before starting to commit, but
there's still a race where the checkpoint thread can start another
checkpoint after the commit finishes.