Add another fix, improving the handling of fast-truncate invariants/states.
The description of the fast-truncate structure invariants that we
originally put in WT-9252 turned out to be missing a case. And there
were two places in the code that weren't handling that case; one is
probably innocuous but the other can cause deleted pages to come back
to life.

Fix those cases and improve the invariant description, being much more
explicit about the possible states. And since this is clearly delicate
and easy to get wrong, lay down a lot more comments in the places
where fast-truncate things happen, and add a couple assertions.

Also, in the interests of clarity, rename:

   __wt_delete_page_rollback -> __wt_delete_page_abort
   It is no longer used by RTS, only by transaction abort, and the
   logic depends on that, so the new name makes things clearer.

   __wt_txn_op_apply_prepare_state -> __wt_txn_op_delete_apply_prepare_state
   It is only for delete ops; this matches other usage.