WT-9168 Assert non-null update passed to update_serial (#7897)
It doesn't make sense to call this with a null update, and in any case it was broken if we did (since it dereferences the update in a couple of spots).

I also had a look at the call sites, there are only two: one in col_modify, and one in row_modify. If either of these manage to pass in a null update, something is very wrong.