WT-9251 Perform a log flush by checkpoint transaction commit
Checkpoint may include the updates from a recent transactions
due to some parallel eviction activity. Ensure that all the log
writes that are performed before metadata gets checkpointed
as part of the checkpoint commit transaction to ensure that all
the log reecords are written to the disk before the checkpoint
is become stable.