Create a new branch for WT-7833 that includes the previous changes.
* Allow concurrent writes to any number of WT_BLOCK handles. When we switch files there can
be any number of writers in flight to the previous file. Because checkpoints are limited to a
single file, all writers must have drained after a subsequent checkpoint, and then we can close
the original WT_BLOCK reference.

* Lock when extending the WT_BLOCK list of cached handles, two threads can race doing the
extension (but not when checking the cached handle list, of course).

* 1000 flushes takes a long time, drop it to 500.

* We need finer-grained control over the objects we drain at checkpoint, checkpointing one
file doesn't mean we can drain the handles for any other file. Create a list of files as part
of switching to a new object, that's the specific list that needs draining.

* Remove verbose output.

* If you checkpoint the current object, then switch objects, and never checkpoint the new object,
you can end up with checkpoint drain chains at database close. It's slightly difficult to handle
because the objects won't be closed in any particular order, assert we're in close because any
other path would be a sign of a serious problem.

* The block cache layer is going to need its own lock, put it in the WT_BLOCK structure for now.

* Don't trust the higher-level locking, confirm it's never wrong.