WT-9251 Perform log flush before the metadata checkpoint (#7910)
The checkpoint may include the updates from a recent transactions
due to some parallel eviction activity. Ensure that all the log
writes that are performed in parallel are flushed to disk before the
metadata gets checkpointed to ensure that all the log records are
written to the disk before the checkpoint becomes stable.