Rework some stuff to reduce the number of states.
 - Kill off most of the special handling of readonly trees. Instead of
hanging onto ft_info.del after instantiation, which creates special
cases all over the place, observe that in the two critical places we
need the information (page read and __wt_btcur_skip_page) we already
unpack the parent's address cell and can refetch the delete
information from there.

 - It is still the case that in readonly trees we can end up with
deleted pages in state WT_REF_DISK. It turns out that teaching
eviction to set the state back to WT_REF_DELETED when discarding an
instantiated page in a readonly tree is a mess, so don't, and instead
continue to handle this case. However, it only manifests in the same
two places so we can get the delete information from the parent
address cell.

 - This means that the ft_info union uses ft_info.del when the state
is WT_REF_DELETED and ft_info.update when the state is WT_REF_MEM, and
otherwise the pointer is always NULL, which makes the union accesses a
lot simpler and less likely to go wrong.

Secondary changes:

 - The WT_ADDR_COPY structure now needs to contain a WT_PAGE_DELETED
and this means it needs to move (otherwise the include order is wrong)
so put it in btree_inline.h where it's used; it's specific to
__wt_ref_addr_copy that lives there so this seems appropriate. Also
resort the structure by use frequency instead of for padding
avoidance, since the latter really doesn't matter here.

 - We no longer check __wt_page_del_active when instantiating, because
it's awkward with the current flow. This means we instantiate with
possibly obsolete timestamps and maybe incur more visibility checks
downstream, but the effect doesn't seem likely to be significant. If
it is, the check can always be restored.