Fix a race between instantiate and internal page reconciliation.
If a deleted page gets instantiated when a checkpoint is already
running, and this happens after the checkpoint would have reconciled
it but before its parent internal page gets reconciled, the internal
page reconciliation finds that it's never been reconciled since it was
modified. This causes it to assume it was modified after the
checkpoint started, so the modifications don't need to be in the
checkpoint and it takes a reference to the preexisting (unmodified)
page image.

Unfortunately, in the case where the _delete_ was before the
checkpoint started, and only the instantiation was after, the deletion
may be needed in the checkpoint, and if so this is wrong and causes
the page to come back from the dead.

At that point it's too late to reconcile the page, so we need to write
out the reference to the original page with a proxy (page-deleted)
cell. This means we need the page_delete information at this point.

To fix this, on instantiation move the page_delete structure into
page->modified, and dispose of it only after the page is next
reconciled. Also add a flag in page->modified to indicate whether the
page is instantiated; since the page_delete structure can be NULL to
mean "the delete is globally visible", using page->modified->page_del
to test if the situation applies doesn't work.

This patch is somewhat preliminary and not yet fully proven, but I'm
pushing it now so it's readily accessible.

Currently, it only does anything with page->modified for read-write
trees on the grounds that the info isn't needed in read-only trees;
but this introduces an extra case that also doesn't really serve any
purpose, so may change.

Also, the new code in __wt_rec_child_modify is just enough like
__rec_child_deleted to want to share it, but just not enough to make
that a nuisance, and tidyups are probably going to happen.