Implement real visible_all for reading checkpoints.
I'd decided it was better to patch/shortcut the call sites (there
aren't that many) rather than complexify the visibility code.

However, it turns out that reading from the history store *requires* a
checkpoint-level visible_all: history store entries can have out of
order stop times, and in particular old values with globally-visible
stop times can hang around and appear out of stop-time order. It's
necessary to skip over these to be able to see a subsequent value with
a zero start time and still-current stop time. It's also necessary to
have a checkpoint-relative visible_all to do it, or with a
sufficiently old checkpoint all history values disappear.

Implement the necessary checks (there are only two places that need to
be changed) and fix the history store read path.

Thanks to Chenhao for explaining what was going on when I was barking
up completely the wrong tree.